refactored code to have helper functions imported in mergeSort.tsx
